requirements.txt
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.1.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blessings==1.7
blinker==1.7.0
boto3==1.26.55
botocore==1.29.55
Brotli==1.1.0
cachelib==0.10.1
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cssselect2==0.7.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.1
dnspython==2.2.1
email-validator==1.3.0
emoji==2.2.0
emojis==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fake-useragent==1.0.1
fastjsonschema==2.16.2
filelock==3.12.3
finance-calculator==0.0.6
Flask==3.0.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Excel==0.0.7
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-Session==0.6.0
Flask-SQLAlchemy==3.0.2
flask_weasyprint==1.1.0
fonttools==4.38.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.1
gspread==5.10.0
gTTS==2.3.0
h11==0.14.0
html5lib==1.1
idna==3.4
instaloader==4.10
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
keras==2.11.0
kiwisolver==1.4.4
lml==0.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
passlib==1.7.4
pdfkit==1.0.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.10.0
playsound==1.2.2
plyer==2.1.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycparser==2.21
pydyf==0.9.0
pyee==8.2.2
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
Pygments==2.13.0
PyMySQL==1.1.0
pyparsing==3.0.9
pyphen==0.14.0
pypng==0.20220715.0
PyQRCode==1.2.1
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pywinpty==2.0.9
pyzmq==24.0.1
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.10.1
segno==1.5.2
selenium==4.6.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speedtest-cli==2.1.3
SQLAlchemy==1.4.45
sqlparse==0.4.4
stack-data==0.6.1
stripe==5.0.0
tabulate==0.9.0
terminado==0.17.0
texttable==1.6.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.24.4
w3lib==2.0.1
wcwidth==0.2.5
weasyprint==61.1
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==3.0.1
wkhtmltopdf==0.2
wsproto==1.2.0
WTForms==3.0.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.10.0
zopfli==0.2.3
